More robust path_concat

+++ include/kioku/filesystem.h
@@ -12,9 +12,9 @@
  * \param[in] destsize Number of bytes available for storage in dest.
  * \param[in] path1 Left side of the path to store.
  * \param[in] path2 Right side of the path to store.
- * \return Number of bytes placed in dest. -1 is returned if any of the pointers are NULL, or if there isn't enough room to store the concatenated path.
+ * \return Number of bytes placed in dest, excluding the NULL byte ('\0') at the end. -1 is returned if any of the pointers are NULL, or if there isn't enough room to store the concatenated path. In such cases the contents of dest are unspecified.
  */
-ssize_t kioku_path_concat(char *dest, size_t destsize, const char *path1, const char *path2);
+int32_t kioku_path_concat(char *dest, size_t destsize, const char *path1, const char *path2);
 
 /**
  * The concept of a user home directory isn't really a cross-platform concept.
